[HUDI-6648] Support building table views from existing files (#9367)
This change allows us to support creating a table from a set of initial files 
that were not created by Hudi. This will allow us to bootstrap tables from 
an existing set of files instead of requiring users to rewrite tables with 
filenames in a format that Hudi expects (outside of the bootstrap feature).

This feature only works if metadata table is enabled.