[HUDI-6736] Fixing rollback completion and commit timeline files removal (#9521)
The purpose of 8849 change is to fix the ordering of rollbacks such that, the completion of rollback instant happens first followed by commits file removal from the timeline.
For eg,
if t5.c.inflight is partially failed, and t6.rb.requested is triggered to rollback.
towards the completion, t6.rb is moved to completed state. and later all t5 commit files are removed from the timeline.
This could lead to dangling commit files (t5) if the process crashes just after moving the t6 rollback to completion. So, 8849 also introduced polling completed rollbacks and ensure we don't trigger another rollback for t5.

But we missed that we already landed 5148 which was addressing a similar issue.
As per 5148, we first need to delete the commit files from timeline (t5) and then transition the rollback to completion (t6.rb). So, even if there is a crash, if we re-attempt t6.rb.requested, it will get to completion w/o any issues (even if t5 is not in the timeline at all).
Hence reverting some of the core changes added as part of 8849. But there are some tests added and so not reverting the entire patch.

---------

Co-authored-by: Jonathan Vexler <=>
Co-authored-by: sivabalan <n.siva.b@gmail.com>