[HUDI-6758] Fixing deducing spurious log blocks due to spark retries (#9611)
- We attempted a fix to avoid reading spurious log blocks on the reader side with #9545.
When I tested the patch end to end, found some gaps. Specifically, the attempt Id we had with taskContextSupplier was not referring to task's attempt number. So, fixing it in this patch. Tested end to test by simulating spark retries and spurious log blocks. Reader is able to detect them and ignore multiple copies of log blocks.