[HUDI-6678] Fix the acquisition of clean&rollback instants to archive (#9416)
The current `getCleanInstantsToArchive` filters clean and rollback instants according to maxInstantsToKeep and minInstantsToKeep, respectively.  There are two disadvantages: (1) If user has only a few rollback instants (not satisfied with the archive), then they will exist forever, even if the commit instant has become very large, which can be very confusing for users; (2) Archiving clean and rollback separately will cause holes in the active timeline.  This commit improves the logic of `getCleanAndRollbackInstantsToArchive` to achieve an active timeline of continuous instants.