[MINOR] Fix default config values if not specified (#9625)
The default values for the configs below are incorrect:

1. hoodie.datasource.write.row.writer.enable
2. hoodie.clustering.preserve.commit.metadata (getPreserveHoodieMetadata)

The default values are not loaded from `#defaultVal` as the configurations are defined in a module-scope that is inaccessible by the current scope. This is why config keys are defined as string here.

This commit fixes these inconsistencies first. Subsequent refactoring might be required to move these config-keys to a scope that is accessible by all other (relevant) modules.

**Note:** The existing test coverage does not cover clustering performed using the RowWriter API. Only RDD API is included as of now. 

Co-authored-by: voon <voonhou.su@shopee.com>