[HUDI-6882] Differentiate between replacecommits in cluster planning (#9755)
Cluster planning will run clustering every n commits. To do this, it gets the previous clustering instant and then finds the number of commits after that. However, it was finding the previous clustering instant just by finding the latest replacecommit. Replacecommit is also used for insert_overwrite. This commit fixes the logic to check the commit metadata to ensure it is a cluster commit.

Co-authored-by: Jonathan Vexler <=>