[HUDI-6873] fix clustering mor (#9774)
Currently during clustering of noncompacted mor filegroups with row writer disabled 
(currently the default for clustering), the records in the base file are applied to the 
log scanner after the log files have been scanned. If they have the same precombine, 
the base file records will be chosen over the log file records. This commit mimics the 
implementation in Iterators.scala to make the behavior consistent.

---------

Co-authored-by: Jonathan Vexler <=>