[HUDI-7053] Fix the filter pushdown logic for file group reader
For MOR table, when position based merging is specified, no filters should be pushed down; otherwise, the filters can be pushed down for key based one.

TESTS:

need to make the demo work to demonstrate that the filter has been pushed down.
need to prove the the pushdown is logically correct for key based merging.