[HUDI-7243] TVF to query hudi timeline through spark sql
A new TVF, `hudi_timeline_query(...)` is added to support querying timeline through spark-sql.
The information displayed is influenced by the 'Commits' command of hudi-cli (with some additional information).

A new relation, `TimelineRelation`, is added to transparently support this functionality. The relation implements
`buildScan(...)` method of `TableScan` trait. It does not support filter or predicate push-down. Column filtering
and predicate evaluation needs to be done by the execution layer. This seems reasonable for the initial implementation
for this tool which is mainly going to be used as a debugging/introspection tool.
The relation defines a fixed schema based on the information present ina Hudi timeline instant.