[HUDI-7261] TVF to query hudi table's filesystem state through spark-sql
Depends on [HUDI-7243]

A new TVF, `hudi_filesystem_view(...)` is added to support querying timeline through spark-sql. The information displayed is influenced by the 'fsview' command of hudi-cli

A new relation, `FileSystemRelation`, is added to transparently support this functionality. The relation implements buildScan(...) method of TableScan trait. It does not support filter or predicate push-down. Column filtering and predicate evaluation needs to be done by the execution layer. This seems reasonable for the initial implementation for this tool which is mainly going to be used as a debugging/introspection tool. The relation defines a fixed schema required to display basic file information of a given hudi table