[HUDI-7294] TVF to query hudi metadata (#10491)
Adds a TVF function to query hudi metadata through spark-sql. Since the metadata is already a MOR table, it simply creates a 'snapshot' on
a MOR relation. Could not find any way to format (or filter) the RDD generated by the MOR snapshot relation. Uploading the PR to get some feedback.

Co-authored-by: Vinaykumar Bhat <vinay@onehouse.ai>