[HUDI-7316] AbstractHoodieLogRecordReader should accept HoodieTableMetaClient in order to reduce occurences of executors making file listing calls when reloading active timeline (#10540)
Summary:
Currently some implementors of  AbstractHoodieLogRecordReader create a HoodieTableMetaClient on construction, which implicitly reloads active timeline, causing a `listStatus` HDFS call. Since these are created in executors, each of the hundreds to thousands of executors will make a `listStatus` call at the same time during a stage. To avoid these redundant calls to HDFS NameNode, AbstractHoodieLogRecordReader and the following implementations have been updated to allow an existing HoodieTableMetaClient to be passed in.
- HoodieUnMergedLogRecordScanner
- HoodieMergedLogRecordScanner
- HoodieMetadataMergedLogRecordReader
As long as the caller passed in a HoodieTableMetaClient with active timeline already loaded, and the implementation doesn't need to re-load the timeline (such as in order to get a more "fresh" timeline) than `listStatus` calls can be avoided in the executor, without causing the logic to be incorrect.

Co-authored-by: Krishen Bhan <bkrishen@uber.com>