[HUDI-7146] Handle duplicate keys in HFile
This PR is borrowed from PR 10211 (https://github.com/apache/hudi/pull/10211).
This lays the groundwork for supporting duplicate keys (non-unique keys) in the HFile reader and writer.
The writer part is fully supported, but the reader side requires some additional support to merge base file and log records (which will come in a subsequent PR).
This PR also fixes some minor issues with the reader which was failing a RecodIndex unit tests.