[HUDI-7357] Introduce generic StorageConfiguration (#10586)
This commit introduces the generic `StorageConfiguration` to store configuration for I/O with `HoodieStorage`. Given there's overhead of reinitializing Hadoop's `Configuration` instance, the approach is to wrap the instance in the `HadoopStorageConfiguration` implementation.  This change will enable us to remove our dependency on Hadoop's `Configuration` class.  When integrated, places using `Configuration` will be replaced by `StorageConfiguration` and the `StorageConfiguration` will be passed around for instantiating `HoodieStorage` (unless Hadoop-based readers need the `Configuration` instance).