[HUDI-7416] Add interface for SourceProfile to be used in StreamSync for reading data (#10736)
Introducing a new class known as SourceProfile which contains details about how the next sync round in StreamSync should be consumed. For eg:

KafkaSourceProfile contains number of events to consume in this sync round.
S3SourceProfile contains the list of files to consume in this sync round
HudiIncrementalSourceProfile contains the beginInstant and endInstant commit times to consume in this sync round.
In future we can add the method for choosing the writeOperationType and indexType as well, for sourceProfile.getSourceSpecificContext() will be used to consume the data from the source.