[HUDI-7472] prevent MDT partitions from getting dropped (#10804)
The functional index creation code-path creates a HudiTable object and gets a metadata writer.
But, this code path (of creating metadata writer) also deletes the existing MDT partitions
iff the write-config does not contain the relevant MDT/index configs. This logic is contained
within HoodieTable::deleteMetadataIndexIfNecessary.

The existing code in HoodieSparkFunctionalIndexClient::create is the entry point for
functional index creation. This creates a custom write-config in
HoodieSparkFunctionalIndexClient::buildWriteConfig which is then used to create a client for
the base table (on which the functional index needs to be created). This PR fixes the
issue noted earlier by adding the relevant MDT partitions config in
HoodieSparkFunctionalIndexClient::buildWriteConfig. A test is also added to ensure that creating
of functional index does not drop the existing MDT partitions.

Co-authored-by: Vinaykumar Bhat <vinay@onehouse.ai>