[HUDI-7512] sort input records for insert operation
Sort the input records based on new configs for INSERT operation.
The records are globally sorted (within an input batch). This allows for
efficient pruning of files (based on column stats) on the query path.

BaseSparkCommitActionExecutor sorts the input records (if the configs
are set) based on partition-path + sort-column. The default sort-column is the
record-key (prepended with partition-path).

A new UpsertSortPartitioner is implemented which assigns a contiguous chunk
of sorted input records to a given bucket (or spark RDD partition).