[HUDI-7569] [RLI] Fix wrong result generated by query
Record level index speeds up queries (when appropriate config properties are enabled)
by pruning files based on metadata's RLI partition entries. The current implementation can
prune files only when the query predicate has a EqalTo (i.e '=') OR In filters/expressions on
the record-key column.

However, the logic to detect if a 'In' query predicate references a record-key column is buggy.
This can result in wrong results when the query predicate is a `In` expression ona  column other than
record-key column.

This PR fixes this problem and adds a unit test for the same.