[HUDI-7559] [1/n] Fix RecordLevelIndexSupport::filterQueryWithRecordKey (#10947)
RecordLevelIndexSupport::filterQueryWithRecordKey() throws a NPE if the EqualTo
query predicate is not of the form `AttributeReference = Literal`. This is because
RecordLevelIndexSupport:::getAttributeLiteralTuple() returns null in such cases which
is then derefercend unconditionally.

This bug was rendering the functional index to not be used even when the query predicate
had spark functions on which functional index is built. Hence these column-stats based functional index
was not pruning files.

This PR makes the following minor changes.
1. Move some methods in RecordLevelIndexSupport into an object to make it static (to aid in unit testing)
2. Fix filterQueryWithRecordKey() by checking for null return values from the call to getAttributeLiteralTuple
3. Add unit tests in TestRecordLevelIndexSupport.scala

Co-authored-by: Vinaykumar Bhat <vinay@onehouse.ai>