[HUDI-6497] Replace FileSystem, Path, and FileStatus usage in hudi-common module (#10591)
This commit makes the changes to replace most `FileSystem`, `Path`, and `FileStatus` usage with `HoodieStorage`, `StoragePath` and `StoragePathInfo` (introduced in #10567, renamed in #10672) in `hudi-common` module, to remove dependency on Hadoop FS abstraction which is not essential to most Hudi core read and write logic.

This commit still keeps using the Hadoop FileSystem-based implementation under the hood.  A follow-up PR will make `HoodieStorage` and I/O implementation pluggable.