Pass local engine context instead of sparkContext while requesting clean plan from metadata table
Summary: Some datasets have huge record_index partition, listing this partition is causing OOM errors. So, if we need to increase executor memory we need to increase it for all the executors. So, by passing local engine context letting the driver do the listing and we can increase just driver memory to handle the OOM error.