[HUDI-7391] HoodieMetadataMetrics should use Metrics instance for metrics registry (#10635)
Currently HoodieMetadataMetrics stores metrics in memory and these metrics are not pushed by the metric reporters. The metric reporters are configured within Metrics instance. List of changes in the PR:

Metrics related classes have been moved from hudi-client-common to hudi-common.
HoodieMetadataMetrics now uses Metrics class so that all the reporters can be supported with it.
Some gaps in configs which are added in HoodieMetadataWriteUtils
Some metrics related apis and functionality has been moved to HoodieMetricsConfig. The HoodieWriteConfig APIs now delegate to HoodieMetricsConfig for the functionality.