[HUDI-7635] Add default block size and openSeekable APIs to HoodieStorage (#11048)
This PR adds `getDefaultBlockSize` and `openSeekable` APIs to
`HoodieStorage` and implements these APIs in `HoodieHadoopStorage`.
The implementation follows the same logic of creating seekable input
stream for log file reading, and `openSeekable` will be used by the log
reading logic.

A few util methods are moved from `FSUtils` and
`HoodieLogFileReader` classes to `HadoopFSUtilsclass`.