[HUDI-8005] Add new lock provider implementation
The change adds a new lock provider which is exactly the same as
existing dynamoDB based one except the way the partition key part is
handled.

Existing one: The parameter is explicitly specified in lock
configuration.
New one: It is implicitly derived from the hudi table base path and hudi
table name. The 2 attributes are available in todays lockConfig, which
contains all props from hudiWriterConfig.

Test:
Unit test and integration test