[HUDI-8208] Fix partition stats bound when compacting or clustering (#12050)
The [min, max] range in column stats or partition stats can keep widening with udpates or deletes, because we simply take min of all mins' and max of maxs' while merging the stats. This can lead to a degenerative case where all partitions qualify for a predicate based on stats, even though actually very few partitions meet the predicate based on actual data. It defeats the purpose of pruning/skipping using stats. To fix this problem, we need to bring the range to a tighter bound. In order to do so, this PR:

- Adds a flag in column stats metadata payload - isTightBound - to indicate whether min/max range is a tighter bound based on latest snapshot or not. It is false by default and set to true during compaction or clustering.
- Adds a config to disable calculating tight bounds. Enabled by default for compaction and clustering.
- To calculate tight bound, we look at the colstats partition for the uncompacted or unclustered files and then merge the colstats with that of the compacted or clustered files. Most of the changes are in HoodieTableMetadataUtil.

---------

Co-authored-by: sivabalan <n.siva.b@gmail.com>