[MINOR] Corrected the creation of final properties when using HUDIStreamer.
The code depends on the existence of DEFAULT_DFS_SOURCE_PROPERTIES file. Overridden properties are specified for HoodieStreamer using --props=<filepath>. The code uses the default value of --props=DEFAULT_DFS_SOURCE_PROPERTIES. Since not everyone would want to use and publish the default properties file, we can pass an empty --props value. This fix ensures that when --props is an empty string we simply use the properties supplied as the argument to the HUDIStreamer class.