[MINOR] Preload the file listing for each partition into the FileSystemView so that executors do not have to perform the listing again.
When config.getBloomIndexPruneByRanges() is enabled, the latest parquet file for each fileID is opened to reads the footers for min/max record keys. Finding the latest parquet file requires listing the partition which will be perform many times. Preloading caches the file listing of the partition into the FileSystemView and is serialized and send to the executors.