Refactor hoodie test table
Revised the test utils so it can work in both table version 6 timeline
layout 1 and table version 8 and timeline layout 2 mode. Previously as
all the metafolder path and the completion instant file names are are
hard coded, so they cannot accommodate based on the given version.

FileCreateUtilsV2 is introduced which takes metaClient as input to solve
this issue. We should use this as opposed to old FileCreateUtilsV2 in
the future.