Conditionally enforce data column type mismatch check
In case of partial update, we may hit cased where the partition column
does not have assignement and we could not find a mapping between source
and target table for the partition field. We will be robust to ignore
the resolution failure when doing the column type check.

Similar for precombined field. Plus if it is not event time ordering,
there is no point of enforcing the column type check.