feat: Introduce a representative file containing the estimated total size of file slice
1. Introduce a representative file containing the estimated total size of file slice

NOTE:
When dealing with `HadoopFsRelation`, Spark merges `PartitionedFile` based on data such as file size.
At present, we directly use the base file or a random log file as the `PartitionedFile` of the `FileSlice`.
As a result, spark cannot accurately use representative data when merging.
Therefore, I estimated the size of the entire `FileSlice` file if it is converted into parqeut file.
Using this data to represent the file slice can provide more accurate data for spark to optimize.

Signed-off-by: TheR1sing3un <chaoyang@apache.org>