[HUDI-9205] Introduce a representative file containing the estimated total size of file slice (#13070)
* feat: Introduce a representative file containing the estimated total size of file slice

NOTE:
When dealing with `HadoopFsRelation`, spark merges `PartitionedFile` based on data such as file size.
At present, we directly use the base file or a random log file as the `PartitionedFile` of the `FileSlice`.
As a result, spark cannot accurately use representative data when merging.

To fix the issue, the size of the entire `FileSlice` estimated as parquet format size is used in the delegate file, which can provide more accurate data to spark for shuffle optimization.

---------

Signed-off-by: TheR1sing3un <chaoyang@apache.org>